shuffle http errors up
