+++ spec/lib/escobar/client_spec.rb
@@ -51,4 +51,22 @@ describe Escobar::Client do
     end
   end
   # rubocop:enable Metrics/LineLength
+
+  it "wraps http exceptions well" do
+    err = ArgumentError.new("Missing some stuff")
+
+    response_body    = { id: "two_factor", message: "Missing 2fa" }.to_json
+    response_headers = { offset: "1" }
+    stub_request(:get, "https://heroku.com").to_return(
+      status: 403, body: response_body, headers: response_headers
+    )
+
+    response = Faraday.get "https://heroku.com"
+
+    exception = Escobar::Client::HTTPError.from_response(err, response)
+    expect(exception.status).to eql(403)
+    expect(exception.body).to eql(
+      "{\"id\":\"two_factor\",\"message\":\"Missing 2fa\"}"
+    )
+  end
 end
